[Rename] mock폴더명을 mocks로 변경
