+++ src/components/organisms/modals/SearchModal.tsx
@@ -5,8 +5,8 @@ import ShortProject from '@/components/molecules/search/ShortProject';
 import Modal from '@/components/organisms/modals/Modal';
 import { ModalProps } from '@/components/organisms/modals/modalProps';
 import Tabs from '@/components/organisms/Tabs';
-import { feedItem } from '@/mock/feedItem';
-import { hubItem } from '@/mock/hubItem';
+import { feedItem } from '@/mocks/feedItem';
+import { hubItem } from '@/mocks/hubItem';
 import { useTabsStore } from '@/store/tabStore';
 import useDebounce from '@/utils/useDebounce';
 import { ChevronRightIcon } from '@heroicons/react/24/outline';
